заготовка, скопировал webserver в пакет hw12
