+++ hw12-WebServer/src/main/java/ru/otus/hw12/WebServerSimpleDemo.java
@@ -1,15 +1,12 @@
-package ru.otus;
+package ru.otus.hw12;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import ru.otus.dao.InMemoryUserDao;
-import ru.otus.dao.UserDao;
-import ru.otus.server.UsersWebServer;
-import ru.otus.server.UsersWebServerWithFilterBasedSecurity;
-import ru.otus.services.TemplateProcessor;
-import ru.otus.services.TemplateProcessorImpl;
-import ru.otus.services.UserAuthService;
-import ru.otus.services.UserAuthServiceImpl;
+import ru.otus.hw12.dao.InMemoryUserDao;
+import ru.otus.hw12.dao.UserDao;
+import ru.otus.hw12.server.UsersWebServer;
+import ru.otus.hw12.server.UsersWebServerSimple;
+import ru.otus.hw12.services.*;
 
 /*
     Полезные для демо ссылки
@@ -23,7 +20,7 @@ import ru.otus.services.UserAuthServiceImpl;
     // REST сервис
     http://localhost:8080/api/user/3
 */
-public class WebServerWithFilterBasedSecurityDemo {
+public class WebServerSimpleDemo {
     private static final int WEB_SERVER_PORT = 8080;
     private static final String TEMPLATES_DIR = "/templates/";
 
@@ -31,10 +28,9 @@ public class WebServerWithFilterBasedSecurityDemo {
         UserDao userDao = new InMemoryUserDao();
         Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
         TemplateProcessor templateProcessor = new TemplateProcessorImpl(TEMPLATES_DIR);
-        UserAuthService authService = new UserAuthServiceImpl(userDao);
 
-        UsersWebServer usersWebServer = new UsersWebServerWithFilterBasedSecurity(WEB_SERVER_PORT,
-                authService, userDao, gson, templateProcessor);
+        UsersWebServer usersWebServer = new UsersWebServerSimple(WEB_SERVER_PORT, userDao,
+                gson, templateProcessor);
 
         usersWebServer.start();
         usersWebServer.join();
