+++ hw12-WebServer/src/main/java/ru/otus/hw12/WebServerWithFilterBasedSecurityDemo.java
@@ -1,12 +1,15 @@
-package ru.otus;
+package ru.otus.hw12;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import ru.otus.dao.InMemoryUserDao;
-import ru.otus.dao.UserDao;
-import ru.otus.server.UsersWebServer;
-import ru.otus.server.UsersWebServerSimple;
-import ru.otus.services.*;
+import ru.otus.hw12.dao.InMemoryUserDao;
+import ru.otus.hw12.dao.UserDao;
+import ru.otus.hw12.server.UsersWebServer;
+import ru.otus.hw12.server.UsersWebServerWithFilterBasedSecurity;
+import ru.otus.hw12.services.TemplateProcessor;
+import ru.otus.hw12.services.TemplateProcessorImpl;
+import ru.otus.hw12.services.UserAuthService;
+import ru.otus.hw12.services.UserAuthServiceImpl;
 
 /*
     Полезные для демо ссылки
@@ -20,7 +23,7 @@ import ru.otus.services.*;
     // REST сервис
     http://localhost:8080/api/user/3
 */
-public class WebServerSimpleDemo {
+public class WebServerWithFilterBasedSecurityDemo {
     private static final int WEB_SERVER_PORT = 8080;
     private static final String TEMPLATES_DIR = "/templates/";
 
@@ -28,9 +31,10 @@ public class WebServerSimpleDemo {
         UserDao userDao = new InMemoryUserDao();
         Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
         TemplateProcessor templateProcessor = new TemplateProcessorImpl(TEMPLATES_DIR);
+        UserAuthService authService = new UserAuthServiceImpl(userDao);
 
-        UsersWebServer usersWebServer = new UsersWebServerSimple(WEB_SERVER_PORT, userDao,
-                gson, templateProcessor);
+        UsersWebServer usersWebServer = new UsersWebServerWithFilterBasedSecurity(WEB_SERVER_PORT,
+                authService, userDao, gson, templateProcessor);
 
         usersWebServer.start();
         usersWebServer.join();
