+++ hw12-WebServer/src/main/java/ru/otus/hw12/helpers/FileSystemHelper.java
@@ -1,4 +1,4 @@
-package ru.otus.helpers;
+package ru.otus.hw12.helpers;
 
 import java.io.File;
 import java.net.URLDecoder;
