+++ hw12-WebServer/src/main/java/ru/otus/hw12/model/User.java
@@ -1,4 +1,4 @@
-package ru.otus.model;
+package ru.otus.hw12.model;
 
 public class User {
