+++ hw12-WebServer/src/main/java/ru/otus/hw12/server/UsersWebServerSimple.java
@@ -1,18 +1,17 @@
-package ru.otus.server;
+package ru.otus.hw12.server;
 
 import com.google.gson.Gson;
-import org.eclipse.jetty.security.*;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.handler.HandlerList;
 import org.eclipse.jetty.server.handler.ResourceHandler;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
-import ru.otus.dao.UserDao;
-import ru.otus.helpers.FileSystemHelper;
-import ru.otus.services.TemplateProcessor;
-import ru.otus.servlet.UsersApiServlet;
-import ru.otus.servlet.UsersServlet;
+import ru.otus.hw12.dao.UserDao;
+import ru.otus.hw12.helpers.FileSystemHelper;
+import ru.otus.hw12.services.TemplateProcessor;
+import ru.otus.hw12.servlet.UsersApiServlet;
+import ru.otus.hw12.servlet.UsersServlet;
 
 public class UsersWebServerSimple implements UsersWebServer {
     private static final String START_PAGE_NAME = "index.html";
