+++ hw12-WebServer/src/main/java/ru/otus/hw12/server/UsersWebServerWithBasicSecurity.java
@@ -1,4 +1,4 @@
-package ru.otus.server;
+package ru.otus.hw12.server;
 
 import com.google.gson.Gson;
 import org.eclipse.jetty.security.ConstraintMapping;
@@ -9,8 +9,8 @@ import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.handler.HandlerList;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.util.security.Constraint;
-import ru.otus.dao.UserDao;
-import ru.otus.services.TemplateProcessor;
+import ru.otus.hw12.dao.UserDao;
+import ru.otus.hw12.services.TemplateProcessor;
 
 import java.util.ArrayList;
 import java.util.Arrays;
