+++ hw12-WebServer/src/main/java/ru/otus/hw12/server/UsersWebServerWithFilterBasedSecurity.java
@@ -1,15 +1,15 @@
-package ru.otus.server;
+package ru.otus.hw12.server;
 
 import com.google.gson.Gson;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.servlet.FilterHolder;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
-import ru.otus.dao.UserDao;
-import ru.otus.services.TemplateProcessor;
-import ru.otus.services.UserAuthService;
-import ru.otus.servlet.AuthorizationFilter;
-import ru.otus.servlet.LoginServlet;
+import ru.otus.hw12.dao.UserDao;
+import ru.otus.hw12.services.TemplateProcessor;
+import ru.otus.hw12.services.UserAuthService;
+import ru.otus.hw12.servlet.AuthorizationFilter;
+import ru.otus.hw12.servlet.LoginServlet;
 
 import java.util.Arrays;
