+++ hw12-WebServer/src/main/java/ru/otus/hw12/services/UserAuthService.java
@@ -1,4 +1,4 @@
-package ru.otus.services;
+package ru.otus.hw12.services;
 
 public interface UserAuthService {
     boolean authenticate(String login, String password);
